int overflow fixed

+++ SteamAccountManager.AvaloniaUI/Common/Utils/TimeConverter.cs
@@ -2,8 +2,8 @@
 {
     public static class TimeConverter
     {
-        public static long ToDays(int minutes) => minutes / Time.DAY_IN_MINUTES;
+        public static long ToDays(long minutes) => minutes / Time.DAY_IN_MINUTES;
 
-        public static long ToHours(int minutes) => minutes / Time.HOUR_IN_MINUTES;
+        public static long ToHours(long minutes) => minutes / Time.HOUR_IN_MINUTES;
     }
 }
